Added start_logs_collection() method
This is to automate the cluster logs collection operation.
Had to add a `get_cluster_nodes()` method too, as the documented `nodes='*'`
parameter for startLogsCollection does not seem to work.

The error message that I get back is always the same, `"nodes":{"needs
to be a list of nodes"}` or something along those lines.
I tried a lot of different combinations, but none worked.

Each node data structure returned by `get_cluster_nodes()` contains a `otpNode`
attribute which is what we need to provide to `start_logs_collection()`.