Increase default client request timeout to 59 seconds
There's been observations of REST calls taking 20+ seconds, for example
the ones relating to backups. Increase the timeout to just under a
minute. Going slightly under the minute mark is a defensive measure, to
maximize the chances of timing out due to an actual Couchbase REST
timeout, and not due to read or idle timeouts in any intermediate
network hop (NAT devices, stateful routers/firewalls, etc). We can
increase the timeout later, but then it should be done together with
adding defensive TCP keep-alive settings.